+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Feat: Collect more information for exceptions collected via `FlutterError.onError` (#538)
 * Feat: Sentry Dart Plugin - Alpha version (#545)
+* Feat: HTTP breadcrumbs have the request & response size if available (#552)
 
 # 6.0.0-beta.3